Remove GameStateManager and SceneManager.
