--- sources/Colonization/AngelScriptBinders/BindGameStateType.hpp
@@ -1,6 +1,0 @@
-#pragma once
-#include <Urho3D/AngelScript/Script.h>
-namespace Colonization
-{
-void BindGameStateType (Urho3D::Script *script);
-}
