+++ sources/Colonization/AngelScriptBinders/BindHostActivity.hpp
@@ -2,5 +2,6 @@
 #include <Urho3D/AngelScript/Script.h>
 namespace Colonization
 {
+void BindGameStateType(Urho3D::Script *script);
 void BindHostActivity (Urho3D::Script *script);
 }
